state: split `dataGetters` into separate files
