+++ client/state/data-getters/get-request-activity-id.js
@@ -1,0 +1,3 @@
+const getRequestActivityId = ( siteId, rewindId ) => `activity-single-${ siteId }-${ rewindId }`;
+
+export default getRequestActivityId;
