+++ client/state/data-getters/request-external-contributors-id.js
@@ -1,0 +1,3 @@
+const requestExternalContributorsId = ( siteId ) => `site-external-contributors-${ siteId }`;
+
+export default requestExternalContributorsId;
